Updated the MPI code to now correctly spread the job capacity among the nodes as opposed to using a hardcoded value of 3.
Fixed an issue relating to MPI segfaulting when using threads. (now using MPI_Init_thread)

Increased how strenuous the pattern finding for Jo Tower is.

Made the decrypting nodes now return the string being decrypted alongside the PIN that was used to decrypt it. This was done in case we wanted to validate the returns.